Initial Code Commit
This is the OpenScad Files that Generate the Codex Wheel using the Liber Primus Gemetria Primus.  It is designed to allow the characters to be rearranged and rotated to make finding patterns and symmetries easier.

Please comment if you have issues with the model or would like to see other changes made.