This is a mess
I'll need to do a lot of work to this but I wanted to share the code so people can see how it was working.

There is some gross hacky stuff going on and I didn't get either file particularly cleaned up.

Anyways, Im going to keep working on this but I wanted to share what I have so far.