Some Progress
I've got the section 1 IoC up to about 1.70 adjusted for 26 characters.

Divinity Within is now coming out cleanly from this algorithm

I think the next word might be "Even" but I now suspect character 17 "F" of the 3 letter word is a skip and spells "FOR".

Ill need to refactor a bit to fully support that - for now feel free to take a look at the work so far.